[TASK] Ext.Direct replacement
This change removes the dependency to the TYPO3.ExtJS package. It replaces
the usage of the global ExtDirect controllers and uses instances of
the HttpClient which should from now on be used for server communication.

The HttpClient handling uses Ember.Deferred. Also the usages of
$.Deferred are replaced by Ember.Deferred.

This change will be the basis for more refactoring in the server side
services and JavaScript code, but is already a good improvement.

Releases: master, 1.1
Change-Id: I471ffd18fd8da706b0dccc4fcdd7aacbfd5dd909
Reviewed-on: https://review.typo3.org/28919
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller