[BUGFIX] Correctly rename site root nodes
When renaming the ``site root node`` via the Sites management module
the site name was incorrectly used for setting the new root node name
instead of the specified node name.

This change fixes this by mapping the new node name only to the site
when it has been changed.
Besides this cleans up SiteController, Site and the Edit form.

Fixes: #53330
Releases: 1.0, master
Change-Id: I34d0e88efb2880897eaabd292852d4770bc22626
Reviewed-on: https://review.typo3.org/26345
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich