[BUGFIX] Secondary inspector is rendered outside viewport in FF
The secondary inspector opens from the right to the left using a
transition, but in Firefox the left positioning is incorrect so it
animates to a position which is outside the viewport.

This patch changes the animation to be done on the left css property
to make it less error prone.

Change-Id: I8ac215e6901a5414a6b66824ea8804915815991b
Fixes: #54401
Releases: master, 1.0
Reviewed-on: https://review.typo3.org/28009
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann