[BUGFIX] Remove obsolete route that might break routing
This removes the frontend route "content with URI suffix and type"
that could break routing in frontend and backend in certain
circumstances.

Background:
This route was meant to support custom rendering formats (similar to
the ``typeNum`` feature in TYPO3 CMS). But apparently this is currently
not used anywhere in the code and the same can be achieved with custom
routes and/or the ``{@format}`` suffix.
Since support for configurable URI suffixes (introduced with
I14c8a98f341fbc312c57fee44cd383271778e57b) it is possible to disable
the ".html" suffix from all Neos routes. But with that change the
route mentioned above incorreclty matches all pages in the second
level (e.g. "foo/bar").

Change-Id: If0d116c979ede9ad0901a63a4995d26766f95eef
Fixes: #55492
Releases: master, 1.0
Reviewed-on: https://review.typo3.org/27184
Reviewed-by: Bastian Waidelich
Tested-by: Bastian Waidelich
Reviewed-by: Karsten Dambekalns
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer