[FEATURE] Localized node routing and consistent use of context path
This change adds a LocalizedFrontendNodeRoutePartHandler that can be
configured for sites using localization. A new strategy for getting
the locales dimension values from a URI segement is added with
ContentDimensionPresetSourceInterface and a Settings based
implementation.

For consistent use of content dimensions in nodes, several parts of Neos
are updated to use the context path consistently.

Controllers and services in Neos are updated to use the new methods
in the ContextInterface instead of directly accessing the node data
through NodeDataRepository.

How to use
==========

In order to use the localized routing, the following needs to be inserted
to Objects.yaml::

    TYPO3\Neos\Routing\FrontendNodeRoutePartHandlerInterface:
      className: TYPO3\Neos\Routing\LocalizedFrontendNodeRoutePartHandler

Furthermore, a "locale" dimension must be configured, e.g. by using the
following Settings.yaml::

    TYPO3:
      TYPO3CR:
        contentDimensions:
          locales:
            default: mul_ZZ

After that, the site must be re-imported from XML.

This must be merged together with the corresponding changes in the TYPO3CR
package.

Change-Id: I497987d1926724cb91143b1009598abddb191f85
Releases: 1.1, master
Reviewed-on: https://review.typo3.org/28826
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller