[BUGFIX] Fix wrong publishable node counts with custom rel="x" markup
The VIE library that is used to parse entities via RDFa returns entities
also for every "rel" attribute that is used to mark a relation to other
entities in RDFa (see http://www.w3.org/TR/rdfa-syntax/#A-rel).

The check for entities outside the "live" workspace did not check if
the entity attributes contain a workspace at all and just returned true
for unknown entities.

Change-Id: Id271d214275845f24a1aa3f29b1579d52092cd60
Releases: 1.0, 1.1, master