[BUGFIX] Tree is updated on changes
The node tree was not correctly updated on
consecutive changes from the inspector.
This change adds the necessary initializations
after reloading the page to fix the issues.

For testing edit for example the node title more
than once in the inspector and the title in the
trees should also be actualized.

Change-Id: I1a9b63a346b34fa2a032d6581276a4f200f2941e
Releases: master, 1.1, 1.0
Reviewed-on: https://review.typo3.org/28781
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller