[FEATURE] allow to set dimension values in URL if not in live workspace
As preparation for a translation user interface, we need the ability
to set dimension values (like "locale") in the URI, such that it can be
set like::

    features@user-admin;locales=de_DE,mul_ZZ.html

This change allows to set the dimension-values in the URI if not in live
workspace.

This is not yet integrated into the LocalizedFrontendNodeRoutePartHandler.

Additionally, removed a leftover conversion from UUID to nodes when generating URIs.

Releases: 1.1, master
Change-Id: If25a1df32c17566ec5cb47bd638b8280ab7b7c12
Reviewed-on: https://review.typo3.org/29156
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller