[BUGFIX] lastVisitedNode should be reset when changing sites
The lastVisistedNode session variable is always pointing to the
node that was last open in frontend to redirect to that page.
If the node is removed this redirect will lead to a "page not
found" error. This can happen for example if something is
changed in sites management. Therefore changes in sites
management should clear this session variable.

Change-Id: I0d91573946cd46c6f8d62a5518f96c4f61a57441
Fixes: NEOS-127
Releases: master, 1.1
(cherry picked from commit aa059625730c6330ee720fc379f462b1bf55535a)