[BUGFIX] NodeSearchService working as before
Due to the fix of case sensitive like queries merged in Flow
(I53cd80e145eb49bcf8251f0045b9f3ec4fd61105) the NodeSearchService
currently returns wrong results as it cannot properly lowercase
the content of the properties blob field.

As we already discussed this cannot be fixed easily so for now
the NodeSearchService will not attempt to search case insensitive.

This also needs to be backported to all branches that will get the
Flow bugfix.

Change-Id: Ie0979f1a9590a125ef8a8032d58b763ea7214cdf
Releases: master, 1.1, 1.0
Fixes: NEOS-267
Reviewed-on: https://review.typo3.org/31667
Reviewed-by: Karsten Dambekalns <karsten@typo3.org>
Tested-by: Karsten Dambekalns <karsten@typo3.org>