[BUGFIX] Site Export should be able to handle broken assets/images
This is a workaround for NEOS-121 which fixes the site export e.g.
with the demo site.

In order to test this, the demo site should be exported directly
after it was imported. This broke without that change.

Change-Id: Iec62f86ed33b2cf528f01a9e229b813c6707b7c2
Releases: master, 1.1
Related: NEOS-121
(cherry picked from commit d772782eee6e8935dbf315294d9de6b42e7279d6)