[TASK] Improve exception handling during rendering
Improves the exception handling by changing the way
exceptions are thrown for content and by rendering
a fallback page if the whole page couldn't be rendered.

The exception handling for content elements is changed so
that it now only outputs a XML comment in the frontend in
production context. When logged in a short message, without
technical detail is displayed in production context.
In development the output is the same as before, consisting
of a exception message and a TypoScript path.

The exception handling for the whole page is improved so
instead of just always displaying a message when broken,
it renders an error page with a styled message. In development
context a technical message is displayed, while it's only outputted
as a XML comment in production context. This makes it possible to
navigate the backend, even though the current page is broken.

Change-Id: I4540b3d4894d374bf03a8f76136049c6051644a4
Resolves: NEOS-344
Releases: master, 1.2