[TASK] Remove deprecated TemplateImplementation
The Neos TemplateImplementation was deprecated even before
version 1.0 and just outputs an error message, therefor cannot
be in productive use. Removing it will fix the possible error
in the AbstractMenuImplementation that imports the
TemplateImplementation from TypoScript which can  lead to a fatal
PHP error as the name is already in use in this namespace.

Change-Id: Ifea6a15069d7349404858639d1d69964269d9734
Releases: master, 1.2
Resolves: NEOS-738