[BUGFIX] Shake effect on failed login broken
A regression was introduced with the update of jQueryUI,
breaking the shake effect shown on failed logins.
This is due to the library no longer having the effects included,
and the path to the library not being adjusted.

Change-Id: Ibd8f41878e882dca2639163cfef031465aa3bf78
Related: Ia286cf80246d21337eb1ae6bd3ad3d2aa22833ab
Releases: master, 1.2
Fixes: NEOS-803