[TASK] Login dialog improvements
Prevents the login dialog from being duplicated if multiple request
were restricted.

Also streamlines the styling with the normal login by moving
the failed login message below the form, introducing a shake effect
on failed login.

Additionally indication of a login is in progress is added to the
normal login as well as the dialog login.

Furthermore fixes the login dialog to use native form validation in
the form of required fields.

Change-Id: If8eef691f63662d572803f33b0fe13db8ddf63f7
Releases: master, 1.2
Fixes: NEOS-599
Resolves: NEOS-815