[BUGFIX] Throw exception when trying to import an invalid Sites.xml file
With this change a more meaningful exception is thrown if a site could
not be imported for unknown reasons.
Besides this slightly improves the ``site:import`` command to give
feedback about the imported site on success.

This also adds some cosmetic and non-functional cleanups in order to
increase readabilty and IDE support:

* Import FQN where applicable
* Reduce nesting in ``SiteImportService::importFromFile()``
* Remove unused import statements
* Whitespace fixes

Change-Id: I4a9fd0731354c3d33f3adb48fcfcd0ed96bbabc8
Releases: master, 1.2
Related: NEOS-885