[BUGFIX] Copying nodes created duplicate uriPathSegment entries
On duplicating a document (page), the uriPathSegment property was simply
copied as well, resulting in the same URI for two distinct documents if
the copy is a sibling of the original nodes.

This change amends the slot used to fill the uriPathSegment so those
duplicates are avoided upon copy and can also no longer be caused by
manually passing a used uriPathSegment.

Change-Id: I2315d30256f7d291204363fdde7e761154cbc7e4
Releases: master, 1.2
Fixes: NEOS-970