[BUGFIX] Resource links breaks across HTTP scheme / domain
If a site is accessed through different HTTP schemes (http/https),
the cached content contains links with the first one accessed that
is used for both. Same goes for different domains, if no domains
have been configured for a site. This leads to errors with
cross-origin resource sharing and unexpected behavior.

Change-Id: I0973d54fc3f243fe50a05c001cc03d2ed9e77212
Releases: 1.2, 1.1
Fixes: NEOS-852