[BUGFIX] Allow editing of documents as content elements
This fixes a regression introduced with
Ia11ac5b6c7bef4c30e005b53f21215d0b17637c1 which resulted in document
nodes not receiving the ``neos-contentelement`` class which made them
inaccessible for the inspector. By introducing a flag to render this
special type of document metadata only if asked for it, the regression
is fixed.

Releases: master, 1.2, 1.1
Change-Id: Ia8fa5228e772b0fee15921e02135c4899df39c0a