[BUGFIX] createNewNode renders invalid uri path segments
This change fixes a problem resulting in invalid uri path segments when
creating new document nodes. When a new document was created, the
NodeController would simply take the entered document title as the
node's uri path segment, even if it contained characters which are not
allowed as part of an URL path.

We now use the TYPO3CR's function for rendering valid node names (like
in the past) which provides some more or less meaningful transliteration.

Change-Id: If21dd5cacdae73aa317371913997fb8754d39875
Releases: master, 1.2
Resolves: NEOS-1079
Related: NEOS-1059